[debugger] Use an external libGME_modded with custom reporting implementation
the external libGME_modded has an implementation-agnostic delivery of memory reports now :)
This commit majorly involves taking the agnostic reports and reconnecting it to the visual implementation
of the debugger. In this case, it was done without modularizing the "report" namespace.. This is OK, not discouraged.

Connecting agnostic reports to a visual implementation is as simple as sub-classing Spc_Report (gme/Spc_Report.h) and
overridding the report() function. BEWARE! you must also "pointlessly" declare an object of your new class type