support "raw" scaling
Using the Logical render scaling feature of SDL is nice. It maintains aspect ratio,
and pre-scales mouse coordinates to appear the same to the program. But, I wanted
to experiment with non-aspect ratio scaling, so this implementation adds the necessary
"raw" scaling factors.

In addition, I've added the backtick key (`) to cycle between the scaling filters
available stock in sdl2. The filters are applicable both in logical scaling, or "raw"
scaling.

Pressing Shift-` will toggle Logical scaling vs. Raw scaling.

Because I actually enjoy the look of the window maximized out of aspect ratio on
my laptop, I'm leaving it on by default. It will eventually be user configurable
in the options menu.