[echo] update song file format and use bitflags
At apu level and song file level, prepare to consolidate  instrument booleans
into a "flags" variable, just to be tidy and space saving.

Note that it's possible to implement these flags for instrument chunks
that are only present in song files (SongMeta chunks), or for instrument chunks that are
part of the STI file format. The one in this commit (echo) is not saved into STI
(Whether echo is enabled for that instrument in your song is
not really an instrument customization since its sound  relies on SongSettings)